WIP Refactor
