--- index.js
@@ -1,5 +1,0 @@
-require('dotenv').config({ silent: true })
-
-const dashListener = require('./lib/util/dash-listener')
-
-dashListener.start()
