--- lib/api/server.js
@@ -1,7 +1,0 @@
-// TODO: Create a small Express server for signing on and signing off
-
-// The idea here is to create a React Native app with some buttons for sending hello/goodbye messages quickly
-// to the Shift Lab Slack instance.
-
-// One challenge to overcome so far:
-// - How do we handle authentication/user identification? Expose a route to register devices? Authenticate with Slack?
