+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "axios": "^0.16.2",
     "dotenv": "^4.0.0",
+    "express": "^4.15.4",
     "node-dash-button": "^0.6.1"
   },
   "devDependencies": {
@@ -16,7 +17,8 @@
     "jest": "^20.0.4",
     "lint-staged": "^4.0.4",
     "prettier": "^1.6.1",
-    "sinon": "^3.2.1"
+    "sinon": "^3.2.1",
+    "supertest": "^3.0.0"
   },
   "scripts": {
     "precommit": "lint-staged",
@@ -25,6 +27,9 @@
     "test": "jest"
   },
   "lint-staged": {
-    "*.js": ["prettier --write --no-semi --single-quote", "git add"]
+    "*.js": [
+      "prettier --write --no-semi --single-quote",
+      "git add"
+    ]
   }
 }
