+++ server.js
@@ -1,0 +1,7 @@
+require('dotenv').config({ silent: true })
+
+const app = require('./lib/app')
+
+app.listen(3000, () => {
+  console.log('Server listening on port 3000')
+})
